COMMERCE-621 Added new method to commerce subscription entry helper
