+++ commerce-api/src/main/java/com/liferay/commerce/subscription/CommerceSubscriptionEntryHelper.java
@@ -16,6 +16,7 @@ package com.liferay.commerce.subscription;
 
 import aQute.bnd.annotation.ProviderType;
 
+import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.model.CommerceSubscriptionEntry;
 import com.liferay.portal.kernel.exception.PortalException;
 
@@ -27,8 +28,11 @@ import java.util.List;
 @ProviderType
 public interface CommerceSubscriptionEntryHelper {
 
+	public void checkCommerceSubscriptions(CommerceOrder commerceOrder)
+		throws PortalException;
+
 	public void renewSubscriptionEntries(
-			List<CommerceSubscriptionEntry> cpSubscriptionEntries)
+			List<CommerceSubscriptionEntry> commerceSubscriptionEntries)
 		throws PortalException;
 
 	public void renewSubscriptionEntry(
